+++ pkg/fn/runtime/regclient_lister.go
@@ -1,0 +1,49 @@
+// Copyright 2026 The kpt Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package runtime
+
+import (
+	"context"
+
+	"github.com/regclient/regclient"
+	ref2 "github.com/regclient/regclient/types/ref"
+)
+
+// RegClientLister is a TagLister using the regclient module to list remote OCI tags.
+type RegClientLister struct {
+	Client *regclient.RegClient
+}
+
+var _ TagLister = &RegClientLister{}
+
+func (l *RegClientLister) Name() string {
+	return "regclient"
+}
+
+func (l *RegClientLister) List(ctx context.Context, image string) ([]string, error) {
+	ref, err := ref2.New(image)
+	if err != nil {
+		return nil, err
+	}
+
+	defer func() { _ = l.Client.Close(ctx, ref) }()
+
+	tagList, err := l.Client.TagList(ctx, ref)
+	if err != nil {
+		return nil, err
+	}
+
+	return tagList.GetTags()
+}
