+++ pkg/lib/runneroptions/runneroptions.go
@@ -16,6 +16,7 @@
 package runneroptions
 
 import (
+	"crypto/tls"
 	"fmt"
 	"net/url"
 	"os"
@@ -66,6 +67,8 @@ type RunnerOptions struct {
 	// ImagePrefix determines the prefix ResolveToImage will use when resolving a
 	// partial image reference to a fully qualified image reference
 	ImagePrefix string
+
+	TLSConfig *tls.Config
 }
 
 func (opts *RunnerOptions) InitDefaults(defaultImagePrefix string) {
